refactor: the last four module headers, in one change (#496)
write_state (11), storage (10), customscan (6) and delete_vector (6) — the 33
remaining declarations with exactly one consumer outside their defining file.

ONE PR rather than four, and the reason is how #546 and #547 were reviewed: by
running checks — every new include used, no moved symbol having more than one
consumer — not by reading declarations. Correctness here is established by script,
so four diffs are no easier to review than one, and one diff collides with #546
and #547 once instead of up to fifteen times. Every pair of module PRs sharing a
consumer conflicts on the #include line, which is what made a six-PR series
expensive.

Not stacked, deliberately: a stacked PR auto-closes when its base is deleted on
merge (#534, this morning). Merge in any order; whichever is second wants a
one-minute rebase of the include lines.

A defect in my splitting tool, found by the count assert and worth recording
because it silently produced a WRONG module assignment rather than an error:

columnar.h:353 is

    extern void PgColumnarEnsureStorageRow(Relation rel);\t/* pre-create ... */

which ends in "*/", not ";". The block scanner ran to the next line ending in ";"
and swallowed the declaration below it, filing PgColumnarReserveRowNumbers — a
storage symbol — under write_state.h. It compiles either way, because a
declaration is visible from whichever header the consumer includes; only the
"matched 9 of 10" assert on the NEXT module caught it. Terminator detection now
strips trailing comments.

I checked whether it reached the open PRs. It did not: #546 and #547 carry
exactly their intended sets, 22 and 12, nothing extra, nothing missing, nothing
left behind in columnar.h. That is luck — neither module's declarations happened
to sit beside one with a trailing comment.

All four verified exact: 11/11, 10/10, 6/6, 6/6, no extra symbol, none missing,
none stranded. columnar.h goes 1045 -> 997 lines and 164 -> 131 externs.

Gate: pg15a, pg16a, pg17a, pg18a, pg19a, make clean between each (#536), zero
warnings and zero errors on all five. native_vecdecode 24/24, harness_selftest
54/54, replication 41/41 on pg18a.
