+++ src/columnar_customscan.c
@@ -26,6 +26,7 @@
  */
 #include "columnar.h"
 
+#include "columnar_customscan.h"
 #include "columnar_reader.h"
 #include <math.h>
