+++ src/columnar_projection.c
@@ -23,6 +23,7 @@
 #include "columnar.h"
 
 #include "columnar_metadata.h"
+#include "columnar_write_state.h"
 #include "access/table.h"
 #include "catalog/pg_type.h"
 #include "funcapi.h"
