+++ src/columnar_reader.c
@@ -14,8 +14,10 @@
  */
 #include "columnar.h"
 
+#include "columnar_delete_vector.h"
 #include "columnar_metadata.h"
 #include "columnar_reader.h"
+#include "columnar_storage.h"
 #include "fmgr.h"
 #include "access/detoast.h"
 #include "access/htup_details.h"
