+++ src/columnar_vacuum.c
@@ -21,6 +21,8 @@
  */
 #include "columnar.h"
 #include "columnar_metadata.h"
+#include "columnar_storage.h"
+#include "columnar_write_state.h"
 #include "columnar_compat.h"
 
 #include "fmgr.h"
